requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
django-redis==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.8.0
oauthlib==3.2.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
